requirements.txt
click==8.1.0
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
SQLAlchemy==1.4.32
Werkzeug==2.1.0
zipp==3.7.0
openpyxl>=3.0.0
dnspython>=2.3.0
pandas>=1.3.0
